requirements.txt
gitdb==4.0.12
GitPython==3.1.44
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.1
rich==14.0.0
smmap==5.0.2
